+++ macos/Classes/Utils.swift
@@ -1,0 +1,5 @@
+struct Constants {
+    static let methodChannelName = "simform_audio_waveforms_plugin/methods"
+    static let audioWaveforms = "AudioWaveforms"
+    static let checkPermission = "checkPermission"
+}